+++ R/shade-package.R
@@ -4,7 +4,6 @@
 #' See [this bug](https://github.com/Kitware/dive/issues/1415) for more details.
 #'
 #' @name shade-package
-#' @aliases shade
 #' @title Shiny Shadow Fixer
 #' @author Sam Woodman \email{sam.woodman@@noaa.gov}
 #'
@@ -16,7 +15,7 @@
 #' @importFrom readr cols read_csv write_csv write_lines
 #' @importFrom stringr str_glue str_pad str_remove str_replace_all str_split_1
 #' @importFrom tidyr replace_na
-#' @importFrom utils head
+#' @importFrom utils head packageVersion
 #'
 #' @keywords package
 "_PACKAGE"